bloom-release: update catkin-check-depends and add gbp-release-build
Now we update catkin-check-depends, so all build/exec can be transformed
from key to package name, and --ignore-src, because these keys can
be found in our catkin workspace. We only need to make sure all outside
keys are resolved to package names.

gpb-release-build: So above debian source-packages are often created
successful, but the build process maybe failed, so you can use this
script to build that release repo again.

But we don't use gpb-release-build in bloom-local-deb, because
we need to filter packages in bloom-local-deb...