cowbuilder: bind-mount fixed location /var/cache/pbuilder/result
We use this fixed location instead of temp-gbp location, in this
way we can do everything in the hook script: scan packages and
add source list to ros-local.list, and apt-get update.

So all we need to do is to put created debs in this fixed location,
and we should clean this dir before build the whole release repo,
so we need the permissions of this /var/cache/pbuilder/result dir.